Update by Apr.26th, 2026
NoirVisor for UEFI can be built on Linux now!
Ported Microsoft-optimized implementation of memcpy, memset, memcmp and strlen into NASM, and removed AVX optimizations for these implementations.
Fixed a few clippy issues.
Translated existing MASM code of NoirVisor into LLVM's module-level inline assembly.

+++ src/xpf_core/asm.rs
@@ -892,7 +892,7 @@ pub mod vt
 	
 	#[inline] pub fn vmwriteptr(field:usize,value:usize)->VmxResult<EmptyUnit>
 	{
-		vmwrite(field,value as usize)
+		vmwrite(field,value)
 	}
 	
 	#[inline] pub fn vmwrite64(field:usize,value:u64)->VmxResult<EmptyUnit>
